error

+++ app/Http/Controllers/ReceipeController.php
@@ -1,12 +1,16 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\Category;
 use App\Receipe;
 use Illuminate\Http\Request;
-
+use App\test;
 class ReceipeController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +18,10 @@ class ReceipeController extends Controller
      */
     public function index()
     {
-        $data = Receipe::all();
+        // dd(app('test'));
+        // session(['key1'=>"value1"]);
+        // dd(session('key1'));
+        $data = Receipe::where('author_id',auth()->id())->get();
         return view("home",compact('data'));
     }
 
@@ -25,7 +32,8 @@ class ReceipeController extends Controller
      */
     public function create()
     {
-         return view('create');
+        $category = Category::all();
+        return view('create',compact('category'));
     }
 
     /**
@@ -36,29 +44,14 @@ class ReceipeController extends Controller
      */
     public function store()
     {
-        // dd(request()->all());
         $validatedData = request()->validate([
         'name' => 'required',
-        'ingredients' => 'required',
+        'ingredient' => 'required',
         'category' => 'required',
         ]);
 
-        // $receipe = new Receipe();
-        // $receipe->name = request()->name;
-        // $receipe->ingredients = request()->ingredients;
-        // $receipe->category = request()->category;
-        // $receipe->save();
-
-        // dd(request()->all());
-        // Receipe::create(request()->all());
-        Receipe::create($validatedData);
-
-        // Receipe::create([
-        //     'name' => request()->name,
-        //     'ingredients' => request()->ingredients,
-        //     'category' => request()->category
-        // ]);
-
+        Receipe::create($validatedData + ['author_id' => auth()->id()]);
+        session()->flash("message",'Receipe has created successfully!');
         return redirect('receipe');
     }
 
@@ -68,8 +61,11 @@ class ReceipeController extends Controller
      * @param  \App\Receipe  $receipe
      * @return \Illuminate\Http\Response
      */
-    public function show(Receipe $receipe)
+    public function show(Receipe $receipe, test $test)
     {
+        // dd(app('test'));
+        // dd($test);
+        $this->authorize('view',$receipe);
         return view('show',compact('receipe'));
     }
 
@@ -81,8 +77,9 @@ class ReceipeController extends Controller
      */
     public function edit(Receipe $receipe)
     {
-        // dd($receipe);
-        return view('edit',compact('receipe'));
+        $this->authorize('view',$receipe);
+        $category = Category::all();
+        return view('edit',compact('receipe','category'));
     }
 
     /**
@@ -94,21 +91,15 @@ class ReceipeController extends Controller
      */
     public function update(Receipe $receipe)
     {
-        
+        $this->authorize('view',$receipe);
         $validatedData = request()->validate([
         'name' => 'required',
-        'ingredients' => 'required',
+        'ingredient' => 'required',
         'category' => 'required',
         ]);
-        // dd($receipe);
-        // $receipe->update(request()->all());
-        $receipe->update($validatedData);
-        // $receipe = Receipe::find($receipe->id);
-        // $receipe->name = request()->name;
-        // $receipe->ingredients = request()->ingredients;
-        // $receipe->category = request()->category;
-        // $receipe->save();
 
+        $receipe->update($validatedData);
+        session()->flash("message",'Receipe has updated successfully!');
         return redirect('receipe');
     }
 
@@ -120,7 +111,9 @@ class ReceipeController extends Controller
      */
     public function destroy(Receipe $receipe)
     {
+        $this->authorize('view',$receipe);
         $receipe->delete();
+        session()->flash("message",'Receipe has deleted successfully!');
         return redirect('receipe');
     }
